requirements.txt
functions-framework==3.*
google-cloud-firestore==2.*
google-cloud-secret-manager==2.*
amadeus==9.*
requests==2.*
pytest==8.*
pytest-mock==3.*
